Airflow/requirements.txt
numpy==1.19.5
pandas==1.1.5
pyarrow==6.0.1
datetime==4.3
apache-airflow==2.0.1
pytrends==4.7.3
